refactor: jdbc 모듈로 클래스 이동
