+++ nextstep-jdbc/build.gradle.kts
@@ -1,3 +1,3 @@
 dependencies {
-
+    implementation("org.apache.commons:commons-dbcp2:2.6.0")
 }
