+++ nextstep-jdbc/src/main/java/nextstep/jdbc/ConnectionManager.java
@@ -1,4 +1,4 @@
-package slipp.support.db;
+package nextstep.jdbc;
 
 import org.apache.commons.dbcp2.BasicDataSource;
