+++ nextstep-jdbc/src/main/java/nextstep/jdbc/PreparedStatementSetter.java
@@ -1,4 +1,4 @@
-package slipp.dao;
+package nextstep.jdbc;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
