+++ slipp/src/main/java/slipp/dao/UserDao.java
@@ -1,5 +1,6 @@
 package slipp.dao;
 
+import nextstep.jdbc.JdbcTemplate;
 import slipp.domain.User;
 
 import java.util.List;
